fix bson schema number converter using int instead of long as integer type

+++ src/bson-schema/converters/number.ts
@@ -18,7 +18,7 @@ export const convertNumber: Converter<ZodNumber, NumberBSONSchema> = (type) => {
   for (const check of type._def.checks) {
     switch (check.kind) {
       case "int":
-        schema.bsonType = "int";
+        schema.bsonType = "long";
         break;
       case "min":
         schema.minimum = check.value;
